+++ packages/block-editor/src/components/media-placeholder/index.js
@@ -215,7 +215,7 @@ export class MediaPlaceholder extends Component {
 				notices={ notices }
 				onClick={ onClick }
 				onDoubleClick={ onDoubleClick }
-				mediaPreview={ mediaPreview }
+				preview={ mediaPreview }
 			>
 				{ content }
 			</Placeholder>